refactor(bookmark): move BookmarkService and BookmarkQueryService to services package; update imports and exception handling
- Move BookmarkService and BookmarkQueryService to pt.estga.bookmark.services
- Update package and imports in related classes
- Replace IllegalStateException and IllegalArgumentException with DuplicateResourceException and ResourceNotFoundException
- Add DuplicateResourceException to shared-web module
- Update BookmarkAutoConfiguration package and related auto-configuration import
- Add shared-web dependency to pom.xml
