+++ bookmark/pom.xml
@@ -24,6 +24,11 @@
             <artifactId>shared</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>pt.estga</groupId>
+            <artifactId>shared-web</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+        </dependency>
          <dependency>
             <groupId>pt.estga</groupId>
             <artifactId>mark</artifactId>
