+++ bookmark/src/main/java/pt/estga/bookmark/BookmarkController.java
@@ -5,6 +5,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+import pt.estga.bookmark.services.BookmarkQueryService;
+import pt.estga.bookmark.services.BookmarkService;
 import pt.estga.shared.enums.TargetType;
 import pt.estga.shared.interfaces.AuthenticatedPrincipal;
