+++ shared-web/src/main/java/pt/estga/sharedweb/exceptions/DuplicateResourceException.java
@@ -1,0 +1,23 @@
+package pt.estga.sharedweb.exceptions;
+
+/**
+ * Exception thrown when attempting to create a resource that already exists.
+ */
+public class DuplicateResourceException extends RuntimeException {
+
+    public DuplicateResourceException() {
+        super();
+    }
+
+    public DuplicateResourceException(String message) {
+        super(message);
+    }
+
+    public DuplicateResourceException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DuplicateResourceException(Throwable cause) {
+        super(cause);
+    }
+}
